+++ EssentailFeed/EssentailFeed/Feed Cache/FeedStore.swift
@@ -1,0 +1,16 @@
+//
+//  FeedStore.swift
+//  EssentailFeed
+//
+//  Created by Vishal Wagh on 01/06/22.
+//
+
+import Foundation
+
+public protocol FeedStore {
+    typealias DeletionCompletion = (Error?) -> Void
+    typealias InsertionCompletion = (Error?) -> Void
+
+    func deletionCacheFeed(completion: @escaping DeletionCompletion)
+    func insert(_ items: [FeedItem],timestamp: Date, completion: @escaping InsertionCompletion)
+}
