+++ EssentailFeed/EssentailFeed/Feed Cache/LocalFeedLoader.swift
@@ -36,12 +36,3 @@ public final class LocalFeedLoader {
     }
 }
 
-public protocol FeedStore {
-    typealias DeletionCompletion = (Error?) -> Void
-    typealias InsertionCompletion = (Error?) -> Void
-
-    func deletionCacheFeed(completion: @escaping DeletionCompletion)
-    func insert(_ items: [FeedItem],timestamp: Date, completion: @escaping InsertionCompletion)
-}
-
-
